requirements.txt
tensorflow==2.4.0
python-telegram-bot
